requirements.txt
chromadb
dateparser
flask
functions-framework
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-firestore
google-genai
PyJWT
pydantic
pytz
rapidfuzz
requests
protobuf<=3.20.1